+++ docs/src/news.md
@@ -8,6 +8,7 @@
     - Function **occdds**: One-class classification (OCC) using DD-Simca.
 
 - Modifications
+    - Improvement of outlierness and occ functions.
     - Code cleaning.